demo_project/pyproject.toml
[project]
name = "demo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "mne>=1.10.1",
    "pygame>=2.6.1",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
    "scikit-learn>=1.7.2",
]
